refactoring helpers and writers and clientprodivers etc. in to single CosmosDBService
